requirements.txt
fastapi
uvicorn
python-docx
pandas
openpyxl
pdfconverter
python-multipart
cloudinary